broke navbar css down to seperate components, continued mobile development
