+++ src/layout/authentication/editProfile/editProfile.css
@@ -114,19 +114,10 @@
 
 /* 360 */
 
-@media (min-width: 375px){
-}
-
-@media (min-width: 411px){
-}
-
 @media (min-width: 768px){
   .edit-prof__user-icon--grid {
     grid-template-columns: repeat(6,100px);
   }
-}
-
-@media (min-width: 1024px){
 }
 
 @media (min-width: 1366px){
@@ -137,7 +128,4 @@
     min-height: 0;
     padding: 1.5rem;
   }
-}
-
-@media (min-width: 1920px){
 }
