+++ src/layout/navbar/components/navControl/navControl.css
@@ -1,6 +1,5 @@
 .nav__collapse-button{
   cursor: pointer;
-  display: flex;
   position: absolute;
   color: black;
   top: 10px;
@@ -36,8 +35,6 @@
 @media (min-width: 1366px){
   .nav__collapse-button.inactive {
     right: -50px;
-    -webkit-transition: right var(--fast-transition-speed) ease-in-out;
-    -o-transition: right var(--fast-transition-speed) ease-in-out;
     transition: right var(--fast-transition-speed) ease-in-out;
   }
 }
