+++ src/layout/navbar/components/navProfile/navProfile.css
@@ -1,0 +1,27 @@
+.nav__profile--wrapper {
+  box-sizing: border-box;
+  display: flex;
+  align-items: flex-end;
+  background: var(--bg-color-light);
+  width: 100%;
+}
+.nav__profile--img {
+  aspect-ratio: 1/1;
+  width: 4em;
+  margin: .6em 0 .6em .6em;
+}
+.nav__profile--username {
+  margin: 0 1rem;
+  color: var(--primary-font);
+  font-weight: bold;
+  font-size: 3rem;
+}
+
+@media (min-width: 1366px){
+  .nav__profile--img {
+    width: 50px;
+  }
+  .nav__profile--username {
+    font-size: 1.2rem;
+  }
+}
