+++ src/utils/authForm/authenticationForm.css
@@ -5,24 +5,17 @@ h2 {
 }
 
 .auth-page-container {
-  display: -webkit-box;
-  display: -ms-flexbox;
   display: flex;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
+  justify-content: center;
   font-family: Arial, Helvetica, sans-serif;
 }
 .auth-form-container {
-  padding: 30px 40px;
+  padding: 2em 1em;
   background-color: var(--bg-color-light);
-  border-radius: 10px;
-  border-style: solid;
-  border-color: var(--dark-red);
-  -webkit-box-shadow: 0 0 15px var(--dark-red);
-          box-shadow: 0 0 15px var(--dark-red);
-  width: 50%;
-  max-width: 500px;
+  border-top: .2em solid var(--dark-red);
+  border-bottom: .2em solid var(--dark-red);
+  box-shadow: 0 0 15px var(--dark-red);
+  width: 100%;
 }
 
 .auth-title {
@@ -60,24 +53,14 @@ h2 {
 }
 
 .auth-form {
-  display: -webkit-box;
-  display: -ms-flexbox;
   display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: column;
-          flex-direction: column;
+  flex-direction: column;
   gap: 20px;
   margin-top: 20px;
 }
 .auth-form-inputs {
-  display: -webkit-box;
-  display: -ms-flexbox;
   display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: column;
-          flex-direction: column;
+  flex-direction: column;
   gap: 8px;
 }
 .auth-form-inputs label {
@@ -95,10 +78,7 @@ h2 {
   font-weight: lighter;
 }
 .auth-form-inputs input:focus {
-  -webkit-box-shadow: 0 0 0 1px white;
-          box-shadow: 0 0 0 1px white;
-  -webkit-transition: 0.15s;
-  -o-transition: 0.15s;
+  box-shadow: 0 0 0 1px white;
   transition: 0.15s;
 }
 
@@ -115,8 +95,6 @@ h2 {
 }
 .auth-button:hover {
   background-color: rgba(255, 0, 0, 0.455);
-  -webkit-transition: 0.3s;
-  -o-transition: 0.3s;
   transition: 0.3s;
 }
 
@@ -130,8 +108,6 @@ h2 {
 }
 .reset-password-link:hover {
   color:white;
-  -webkit-transition: 0.3s;
-  -o-transition: 0.3s;
   transition: 0.3s;
 }
 
@@ -141,12 +117,8 @@ h2 {
   font-size: 1.2rem;
 }
 .auth-alternate-group {
-  display: -webkit-box;
-  display: -ms-flexbox;
   display: flex;
-  -webkit-box-pack: justify;
-      -ms-flex-pack: justify;
-          justify-content: space-between;
+  justify-content: space-between;
 }
 .auth-alternate-link {
   text-decoration: none;
@@ -155,8 +127,6 @@ h2 {
 }
 .auth-alternate-link:hover {
   color:white;
-  -webkit-transition: 0.3s;
-  -o-transition: 0.3s;
   transition: 0.3s;
 }
 
@@ -176,31 +146,16 @@ h2 {
 }
 .guest-sign-in__button:hover {
   color:white;
-  -webkit-transition: 0.3s;
-  -o-transition: 0.3s;
   transition: 0.3s;
 }
 
-@media (min-width: 600px){
+/* 360 */
 
-}
-
-@media (min-width: 900px){
-
-}
-
-@media (min-width: 1200px){
-  
-}
-
-@media (min-width: 1400px){
-  
-}
-
-@media (min-width: 1800px){
-  
-}
-
-@media (min-width: 1900px){
-  
+@media (min-width: 768px){
+  .auth-form-container {
+    width: 60%;
+    border: .2em solid var(--dark-red);
+    border-radius: 10px;
+    max-width: 500px;
+  }
 }
